Refatoração: validação de ordens movida para o caso de uso
A lógica de validação de ordens foi transferida da entidade
`Account` para o caso de uso `PlaceOrderUseCase`, centralizando
a responsabilidade e alinhando-se ao princípio de separação
de responsabilidades.

Principais mudanças:
- Validações de existência de ativos e saldo insuficiente
  agora estão no método `PlaceOrder`, com lançamento de
  exceções específicas (`EntityNotFoundException` e
  `InsufficientBalanceException`).
- Métodos de validação removidos de `Account`, simplificando
  o método `AddOrder`.
- Testes ajustados: cenários de validação movidos para
  `PlaceOrderTests.cs` com uso de `FluentAssertions` para
  maior clareza.

Essa refatoração melhora a coesão do código e mantém a
cobertura de testes para cenários críticos.
