+++ Domain/Entities/Account.cs
@@ -58,23 +58,6 @@ namespace Domain.Entities
 
         public void AddOrder(Order order)
         {
-            var assetName = order.Side?.ToUpper() != "BUY"
-                ? order.Market?.Split("/")[0].ToUpper()
-                : order.Market?.Split("/")[1].ToUpper();
-
-            var asset = _assets
-                .FirstOrDefault(asset => asset.AssetName?.ToUpper() == assetName);
-
-            if (asset is null)
-            {
-                throw new EntityNotFoundException($"Asset {assetName} not found");
-            }
-
-            if (asset.Balance < order.Quantity)
-            {
-                throw new InsufficientBalanceException($"Insufficient balance for asset {assetName}");
-            }
-
             _orders.Add(order);
         }
     }
